Unsuccessful attempt to move KansasLava to SizedTypes based on TypeNats
The GHC TypeNats extension looks like a promising approach for
handling type level computations which we really use in KansasLava.

I attempted to re-code the SizedTypes module using TypeNats;
but converting KansasLava here to use it has not been successful.

See more details on the NatsNotReady branch of SizedTypes.

This branch is on hold, waiting for likely enhancements in future GHC releases.